refactor: simplify PDF viewer and text selection logic
- Extract PDF viewer and text selection logic into usePdfViewer hook
- Streamline ScriptContent component by removing redundant state management
- Improve sidebar interaction with selected text
- Add console log for debugging selected text
